requirements.txt
pyyaml
langchain-community
langchain-huggingface
chromadb
flask
flask-cors